Syncing merge up to Assignment 4

+++ Question/src/QuestionService.js
@@ -73,6 +73,10 @@ const editQuestion = async (id, title, complexity, description, tags) => {
 };
 
 const deleteQuestion = async (id) => {
+  if (!id) {
+    throw { status: 410, message: 'No question id is given' };
+  }
+
   try {
     await questionRepository.deleteQuestion(id);
   } catch (err) {
